Client Encryption: Relocate Stream corruption-fix tests into existing suites
Distribute the StreamProcessorDataCorruptionFixTests grab-bag into the
topically-appropriate existing test files and delete it:
- round-trip JSON-fidelity + nested-_ei guard -> StreamProcessorDecryptorTests
- out-of-range integer throw -> StreamProcessorEncryptorTests
- decrypt input-dispose -> SystemTextJsonStreamAdapterTests
- quoted _ef tolerance -> EncryptionPropertiesStreamReaderTests

Reuse each target's existing helpers; coverage unchanged (net8 suite 709/709).
